+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Update to 1.19
+- Added armor fade-out after combat.
+- Fixed some minor edge-cases with capes and combat.